requirements.txt
gradio
opencv-python
numpy>2
ffmpeg-python
scikit-learn
kneed
torch
torchvision
matplotlib
pexels-api-py
random